refactor: 파일 구조 변경, 반경 최대 50000으로 변경
