+++ src/main/java/com/backend/arthere/arts/application/ArtsService.java
@@ -2,7 +2,11 @@ package com.backend.arthere.arts.application;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.backend.arthere.arts.domain.ArtsRepository;
-import com.backend.arthere.arts.dto.*;
+import com.backend.arthere.arts.dto.request.ArtImageByAddressRequest;
+import com.backend.arthere.arts.dto.request.ArtImageByArtNameRequest;
+import com.backend.arthere.arts.dto.request.ArtImageByLocationRequest;
+import com.backend.arthere.arts.dto.request.ArtImageByRevisionDateRequest;
+import com.backend.arthere.arts.dto.response.*;
 import com.backend.arthere.arts.util.LocationUtils;
 import com.backend.arthere.image.util.PresignedURLUtils;
 import lombok.RequiredArgsConstructor;
