+++ src/main/java/com/backend/arthere/arts/domain/ArtsCustomRepository.java
@@ -1,6 +1,11 @@
 package com.backend.arthere.arts.domain;
 
-import com.backend.arthere.arts.dto.*;
+import com.backend.arthere.arts.dto.request.ArtImageByAddressRequest;
+import com.backend.arthere.arts.dto.request.ArtImageByArtNameRequest;
+import com.backend.arthere.arts.dto.request.ArtImageByRevisionDateRequest;
+import com.backend.arthere.arts.dto.response.ArtImageByLocationResponse;
+import com.backend.arthere.arts.dto.response.ArtImageResponse;
+import com.backend.arthere.arts.dto.response.LocationRangeResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
