+++ src/main/java/com/backend/arthere/arts/dto/request/ArtImageByAddressRequest.java
@@ -1,4 +1,4 @@
-package com.backend.arthere.arts.dto;
+package com.backend.arthere.arts.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,23 +10,23 @@ import java.io.Serializable;
 
 @Getter
 @NoArgsConstructor
-public class ArtImageByArtNameRequest implements Serializable {
+public class ArtImageByAddressRequest implements Serializable {
 
     private Long idx;
 
     @NotBlank(message = "검색어를 입력하세요.")
-    private String name;
+    private String query;
 
     @Min(value = 1, message = "limit은 1 이상 10 이하입니다.")
     @Max(value = 10, message = "limit은 1 이상 10 이하입니다.")
-    private Integer limit;
+    private int limit;
 
     public void setIdx(final String idx) {
         this.idx = Long.parseLong(idx);
     }
 
-    public void setName(final String name) {
-        this.name = name;
+    public void setQuery(final String query) {
+        this.query = query;
     }
 
     public void setLimit(final String limit) {
