+++ src/main/java/com/backend/arthere/arts/dto/request/ArtImageByArtNameRequest.java
@@ -1,34 +1,32 @@
-package com.backend.arthere.arts.dto;
+package com.backend.arthere.arts.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
+import javax.validation.constraints.NotBlank;
 import java.io.Serializable;
-import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class ArtImageByRevisionDateRequest implements Serializable {
+public class ArtImageByArtNameRequest implements Serializable {
 
-    private LocalDateTime date;
-
-    @Min(value=1, message = "idx는 1이상입니다.")
     private Long idx;
+
+    @NotBlank(message = "검색어를 입력하세요.")
+    private String name;
 
     @Min(value = 1, message = "limit은 1 이상 10 이하입니다.")
     @Max(value = 10, message = "limit은 1 이상 10 이하입니다.")
-    @NotNull(message = "limit는 필수 값 입니다.")
     private Integer limit;
-
-    public void setDate(final String date) {
-        this.date = LocalDateTime.parse(date);
-    }
 
     public void setIdx(final String idx) {
         this.idx = Long.parseLong(idx);
+    }
+
+    public void setName(final String name) {
+        this.name = name;
     }
 
     public void setLimit(final String limit) {
