+++ src/main/java/com/backend/arthere/arts/dto/request/ArtImageByRevisionDateRequest.java
@@ -1,32 +1,34 @@
-package com.backend.arthere.arts.dto;
+package com.backend.arthere.arts.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import java.io.Serializable;
+import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class ArtImageByAddressRequest implements Serializable {
+public class ArtImageByRevisionDateRequest implements Serializable {
 
+    private LocalDateTime date;
+
+    @Min(value=1, message = "idx는 1이상입니다.")
     private Long idx;
-
-    @NotBlank(message = "검색어를 입력하세요.")
-    private String query;
 
     @Min(value = 1, message = "limit은 1 이상 10 이하입니다.")
     @Max(value = 10, message = "limit은 1 이상 10 이하입니다.")
-    private int limit;
+    @NotNull(message = "limit는 필수 값 입니다.")
+    private Integer limit;
+
+    public void setDate(final String date) {
+        this.date = LocalDateTime.parse(date);
+    }
 
     public void setIdx(final String idx) {
         this.idx = Long.parseLong(idx);
-    }
-
-    public void setQuery(final String query) {
-        this.query = query;
     }
 
     public void setLimit(final String limit) {
